requirements.txt

---

# 📄 `requirements.txt`
```txt
torch>=2.0.0
torchvision>=0.15.0
torchmetrics>=0.11.0
mlflow>=2.5.0
optuna>=3.1.0
numpy>=1.23.0
pandas>=1.5.0
scikit-learn>=1.2.0
matplotlib>=3.6.0
Pillow>=9.0.0
torchsampler>=0.1.2
